requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
urllib3==2.5.0
unidecode
pandas
pyarrow

# Scraper dependencies (para Python 3.10+)
selenium
pandas
boto3
python-dotenv